rename: userInfoStore 파일 위치 변경 (store -> entities/userInfo/model) global.d.ts (types -> shared/model)
