+++ app/onboarding/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { BaseButton } from '@/shared/ui'
-import useUserInfoStore from '@/store/userInfoStore'
+import useUserInfoStore from '@/entities/userInfo/model/userInfoStore'
 
 import Image from 'next/image'
 import Link from 'next/link'
