+++ src/entities/userInfo/hooks/useRefetchUserInfo.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import { getUserInfo } from '../api/userApi'
-import useUserInfoStore from '@/store/userInfoStore'
+import useUserInfoStore from '@/entities/userInfo/model/userInfoStore'
 
 export const useRefetchUserInfo = () => {
   const { setUserInfo } = useUserInfoStore()
