+++ src/shared/model/global.d.ts
