+++ src/views/connect/ui/ConnectStep.tsx
@@ -9,7 +9,7 @@ import {
 import { shareWithKakao } from '@/features/share'
 import { useToast } from '@/shared/lib'
 import { BaseButton, DateButton, ProgressBar, TextButton } from '@/shared/ui'
-import useUserInfoStore from '@/store/userInfoStore'
+import useUserInfoStore from '@/entities/userInfo/model/userInfoStore'
 import { useMutation } from '@tanstack/react-query'
 import { AxiosError, isAxiosError } from 'axios'
 import { format } from 'date-fns'
