+++ src/views/onboarding/ui/page.tsx
@@ -2,7 +2,7 @@
 import { useRefetchCoupleInfo } from '@/entities/couple'
 import { useRefetchUserInfo } from '@/entities/userInfo/hooks/useRefetchUserInfo'
 import { BaseButton } from '@/shared/ui'
-import useUserInfoStore from '@/store/userInfoStore'
+import useUserInfoStore from '@/entities/userInfo/model/userInfoStore'
 
 import Image from 'next/image'
 import Link from 'next/link'
